Implement delete_multi_no_reply API for ASCII protocol
